+++ VpnHood.Client/Diagnosing/DiagnoseTestType.cs
@@ -1,0 +1,15 @@
+﻿namespace VpnHood.Client.Diagnosing
+{
+    public enum DiagnoseTestType
+    {
+        InternetPing,
+        InternetUdp,
+        InternetHttp,
+        VpnServerPing,
+        VpnConnect,
+        VpnInternetPing,
+        VpnInternetUdp,
+        VpnInternetHttp
+    }
+
+}